refactor: move integration_tests/ and unit_tests/ under tests_legacy/
Renames integration_tests/ -> tests_legacy/integration/ and
unit_tests/ -> tests_legacy/unit/, bumping relative imports one level
deeper (../src -> ../../src). Frees up the top-level tests/ namespace
for new, actively-maintained suites (e.g. tests/kusama/), matching
registry's layout, without disturbing this existing set.
