+++ tests_legacy/integration/start_downlaod_in_browser.html
@@ -10,7 +10,7 @@
     <button id="downloadButtonProof">Download Proof Files</button>
 
     <script type="module">
-      import { createBlueprint, Proof } from "../dist/index.js";
+      import { createBlueprint, Proof } from "../../dist/index.js";
 
       function getBlueprintProps(title = "Twitter", slug = "zkemail/twitter") {
         return {
