+++ tests_legacy/integration/submitBlueprint.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe, beforeAll, afterAll } from "bun:test";
-import zkeSdk from "../src";
+import zkeSdk from "../../src";
 
 const sdk = zkeSdk({
   baseUrl: "https://staging-conductor.zk.email",
